:recycle:[Refactor] 전역예외처리 패키지 정리
