+++ src/main/java/com/mathfusion/domain/user/exception/UserException.java
@@ -1,7 +1,7 @@
 package com.mathfusion.domain.user.exception;
 
-import com.mathfusion.domain.exception.ErrorStatus;
-import com.mathfusion.domain.exception.GeneralException;
+import com.mathfusion.global.apiPayload.code.status.ErrorStatus;
+import com.mathfusion.global.apiPayload.exception.GeneralException;
 
 public class UserException extends GeneralException {
     public UserException(ErrorStatus errorStatus){
