+++ src/main/java/com/mathfusion/global/apiPayload/exception/GeneralException.java
@@ -1,23 +1,20 @@
 package com.mathfusion.global.apiPayload.exception;
+import com.mathfusion.global.apiPayload.code.status.ErrorStatus;
+
+//공통 예외
+//각 도메인별 예외가 상속받는 클래스
+//Errorstatus 받아서 예외 만들 수 있도록
 
 
-import com.mathfusion.global.apiPayload.code.BaseErrorCode;
-import com.mathfusion.global.apiPayload.code.ErrorReasonDTO;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-
-@Getter
-@AllArgsConstructor
 public class GeneralException extends RuntimeException {
+    private final ErrorStatus errorStatus;
 
-    private BaseErrorCode code;
-
-    public ErrorReasonDTO getErrorReason() {
-        return this.code.getReason();
+    public GeneralException(ErrorStatus errorStatus){
+        super(errorStatus.getMessage());
+        this.errorStatus = errorStatus;
     }
 
-    public ErrorReasonDTO getErrorReasonHttpStatus(){
-        return this.code.getReasonHttpStatus();
+    public ErrorStatus getErrorStatus(){
+        return errorStatus;
     }
 }
